[Pdm2-11] Rename and rearrange Public directories to better match their purpose.
