+++ src/app/(AuthHub)/Home/screens/HomeScreen.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import HomeContent from "../components/HomeContent";
-import Watermark from "../../../common/components/Watermark";
-import Background from "../../../common/components/Background";
+import Watermark from "../../../../common/components/Watermark";
+import Background from "../../../../common/components/Background";
 import { View } from "react-native";
 
 export default function Home() {
