+++ src/app/(AuthHub)/Login/components/LoginContent.tsx
@@ -1,6 +1,6 @@
 import { View } from "react-native";
 import React from "react";
-import Logo from "../../../common/components/Logo";
+import Logo from "../../../../common/components/Logo";
 import Footer from "./Footer";
 import LoginForm from "./LoginForm";
