+++ src/app/(AuthHub)/Login/screens/LoginScreen.tsx
@@ -1,7 +1,7 @@
 import { View } from "react-native";
 import React from "react";
-import Background from "../../../common/components/Background";
-import Watermark from "../../../common/components/Watermark";
+import Background from "../../../../common/components/Background";
+import Watermark from "../../../../common/components/Watermark";
 import LoginContent from "../components/LoginContent";
 
 export default function Login() {
